+++ jquery.nanogallery.js
@@ -25,7 +25,7 @@
 nanoGALLERY v5.9.0 release notes.
 
 ##### New options
-- **albumMax**: Maxum number of albums to display (Flicrk, Picasa/Google+, nanoPhotosProvider) (0=disabled).  
+- **albumMax**: Maximum number of albums to display (Flickr, Picasa/Google+, nanoPhotosProvider) (0=disabled).  
   *integer; Default: 0*  
 - **galleryEnableKeyboard**: Enables keyboard navigation beween albums.  
   *boolean; Default: false*