+++ changelog.md
@@ -39,7 +39,7 @@ v5.9.0
   (credits: Victor A. Banuelos - https://github.com/vbanuelos)  
   
 ##### Misc
-- fixed support of `nanoPhotosProvider` (https://github.com/Kris-B/nanoPhotosProvider)  
+- fixed `nanoPhotosProvider` compatibility issue (https://github.com/Kris-B/nanoPhotosProvider)  
 - removed option 'jsonCharset'  
 - bugfix: issue #86 - error on hoverOut n[r].toFixed is undefined (credits: Andrea Grassi - https://github.com/andrea-sdl)  
 - bugfix: issue #78 - NGTweenable undefined when loading in a page using AMD and requirejs (credits: Jeff Mills - https://github.com/jefftmills)